背景などデザインを大幅に修正
