+++ src/app/chat-with-omochi/page.jsx
@@ -28,33 +28,16 @@ export default function RomanceBot() {
   };
 
   return (
-    <div className="container mx-auto text-center py-8 px-4 sm:px-6 md:px-8 lg:px-12">
-      <h1 className="text-3xl font-bold mb-4 sm:text-2xl md:text-3xl">恋愛マスターおもちの相談室</h1>
-      <div className="chat-container bg-gray-100 p-4 rounded shadow-md mb-4">
+    <div className="min-h-screen bg-[rgba(255,255,255,0.5)] flex flex-col items-center justify-center p-4">
+      <h1 className="text-4xl font-extrabold text-gray-800 mb-8">恋愛マスターおもちの相談室</h1>
         {chatHistory.map((chat, index) => (
           <div
             key={index}
-            className={`chat ${chat.role === 'user' ? 'chat-end' : 'chat-start'} mb-2`}
+            className={`flex items-start ${chat.role === 'user' ? 'justify-end' : 'justify-start'} mb-4`}
           >
-            {chat.role === 'user' && (
-              <div className="chat-image avatar">
-                <div className="w-10 h-10 sm:w-12 sm:h-12 md:w-16 md:h-16 rounded-full">
-                  <Image
-                    alt="User avatar"
-                    src="/bunny.png" // 画像のパス
-                    layout="responsive"
-                    width={100} // 画像の幅
-                    height={100} // 画像の高さ
-                  />
-                </div>
-              </div>
-            )}
-            <div className="chat-bubble bg-white p-3 rounded shadow-md text-left max-w-xs mx-auto sm:max-w-sm md:max-w-md">
-              {chat.content}
-            </div>
             {chat.role === 'rabbit' && (
-              <div className="chat-image avatar">
-                <div className="w-10 h-10 sm:w-12 sm:h-12 md:w-16 md:h-16 rounded-full">
+              <div className="chat-image avatar mr-4">
+                <div className="w-10 h-10 sm:w-12 sm:h-12 rounded-full overflow-hidden">
                   <Image
                     alt="Rabbit avatar"
                     src="/bunny.png" // 画像のパス
@@ -65,20 +48,40 @@ export default function RomanceBot() {
                 </div>
               </div>
             )}
+            <div
+              className={`chat-bubble p-4 rounded-2xl shadow-lg text-left max-w-xs sm:max-w-sm md:max-w-md ${
+                chat.role === 'user' ? 'bg-gray-600 text-white' : 'bg-white text-gray-800'
+              }`}
+            >
+              {chat.content}
+            </div>
+            {chat.role === 'user' && (
+              <div className="chat-image avatar ml-4">
+                <div className="w-10 h-10 sm:w-12 sm:h-12 rounded-full overflow-hidden">
+                  <Image
+                    alt="User avatar"
+                    src="/bunny.png" // 画像のパス
+                    layout="responsive"
+                    width={100} // 画像の幅
+                    height={100} // 画像の高さ
+                  />
+                </div>
+              </div>
+            )}
           </div>
         ))}
-      </div>
-      <div className="mt-4 flex flex-col sm:flex-row sm:gap-4">
+
+      <div className="mt-6 flex flex-col sm:flex-row sm:gap-4 w-full max-w-3xl">
         <input
           type="text"
           value={message}
           onChange={(e) => setMessage(e.target.value)}
-          className="border px-4 py-2 w-full mb-2 sm:mb-0 sm:w-3/4"
+          className="border border-gray-300 px-4 py-2 w-full mb-2 sm:mb-0 sm:w-3/4 rounded-xl focus:ring-2 focus:ring-custom-color outline-none transition-all duration-200"
           placeholder="恋愛相談を書いてください..."
         />
         <button
           onClick={sendMessage}
-          className="btn px-4 py-2 w-full sm:w-1/4 bg-blue-500 text-white rounded shadow-md hover:bg-blue-600 transition"
+          className="btn px-4 py-2 w-full sm:w-1/4 rounded-xl bg-[#cb8e7e] text-white shadow-lg hover:bg-[#b77a6b] transition-transform transform hover:scale-105 duration-200"
         >
           送信
         </button>
