+++ src/app/globals.css
@@ -26,6 +26,10 @@ body {
   background: rgb(var(--background-rgb)); /* 背景色を設定 */
   padding: 0 1rem; /* サイドのパディングを追加 */
   box-sizing: border-box; /* ボックスサイズにパディングを含める */
+  /* 背景画像を追加 */
+  background-image: url('/toppage_bg.png');
+  background-size: cover; /* 画像が全体を覆うように設定 */
+  background-position: center; /* 画像の中心を表示 */
 }
 
 .container {
@@ -56,15 +60,6 @@ body {
   .text-balance {
     text-wrap: balance;
   }
-}
-
-.chat-bubble {
-  background-color: rgb(238, 203, 213) !important;
-  color: white !important;
-  border-radius: 1rem; /* チャットバブルの角を丸くする */
-  padding: 1rem; /* 内側のパディングを追加 */
-  max-width: 90%; /* 最大幅を90%に設定 */
-  margin: 0 auto; /* 中央に配置 */
 }
 
 /* レスポンシブ対応 */
